Split ObjTime into time module
